# apiserver/model_filter_rule.go
/*
 * MS Graph App
 *
 * API to access and configure the MS Graph App
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apiserver

// FilterRule - Asset selection rule. Possible parameters are defined in app's documentation.
type FilterRule struct {
	Parameter string `json:"parameter,omitempty"`

	Regex string `json:"regex,omitempty"`
}

// AssertFilterRuleRequired checks if the required fields are not zero-ed
func AssertFilterRuleRequired(obj FilterRule) error {
	return nil
}

// AssertRecurseFilterRuleRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of FilterRule (e.g. [][]FilterRule), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseFilterRuleRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aFilterRule, ok := obj.(FilterRule)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertFilterRuleRequired(aFilterRule)
	})
}